requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
httpx>=0.27.0
pytest>=8.2.0
pywebview>=5.1
mlx-lm>=0.20.6; platform_system == "Darwin"
